Add Idea Editing
Fix #17
Add summary_lock to idea schema. This will allow admins to lock and
unlock the summary after it has been locked by the first branch or vote.
Add an appended summary array to allow the user to append to the summary
after it has been locked.
Also add a "force" parameter to allow the admin to directly modify the
summary without locking or unlocking it.
Allow for the editing of tags.