Change how voting works
Instead of traversing the database, which is next to impossible because
it's asynchronous, store all users capable of voting on an idea in the
idea itself.

Keep all votes as false initially, and modify them as needed.
TODO
When a new branch is merged, update the votes upstream of that branch.